Created the Franken-file Cards.cpp/h
Commented out the actual file structure because we aren't allowed to have nice things apparently
