+++ warzone/Card.cpp
@@ -1,116 +1,116 @@
-#include "Card.h"
-
-namespace Cards {
-
-	Card::Card() {
-		this->name = new string("Basic Card");
-	}
-
-	Card::~Card()
-	{
-		delete name;
-	}
-	
-	void Card::play()
-	{
-		cout << "this is a basic Card" <<endl;
-	}
-
-	string Card::getName()
-	{
-		return *name;
-	}
-
-	void Card::print()
-	{
-		cout << this->name;
-	}
-
-
-	SpyCard::SpyCard()
-	{
-		this->name= new std:: string("Spy Card");
-	}
-
-	SpyCard::~SpyCard()
-	{
-	}
-
-	void SpyCard::play()
-	{
-		cout << "Playing SpyCard" << endl;
-	}
-
-	BombCard::BombCard()
-	{
-		this->name = new std::string("BombCard");
-	}
-
-	BombCard::~BombCard()
-	{
-	}
-
-	void BombCard::play()
-	{
-		cout << "Playing BombCard" << endl;
-	}
-
-	ReinforcementCard::ReinforcementCard()
-	{
-		name = new std::string("ReinforcementCard");
-	}
-
-	ReinforcementCard::~ReinforcementCard()
-	{
-	}
-
-	void ReinforcementCard::play()
-	{
-		cout << "Playing ReinforcementCard" << endl;
-	}
-
-	BlockadeCard::BlockadeCard()
-	{
-		name = new std::string("BlockadeCard");
-	}
-
-	BlockadeCard::~BlockadeCard()
-	{
-		
-	}
-
-	void BlockadeCard::play()
-	{
-		cout << "Playing BlockadeCard" << endl;
-	}
-		
-	AirliftCard::AirliftCard()
-	{
-		name = new std::string("AirliftCard");
-	}
-
-	AirliftCard::~AirliftCard()
-	{
-		
-	}
-
-	void AirliftCard::play()
-	{
-		cout << "Playing AirliftCard" << endl;
-	}
-
-	DiplomacyCard::DiplomacyCard()
-	{
-		name = new std::string("DiplomacyCard");
-	}
-
-	DiplomacyCard::~DiplomacyCard()
-	{
-		
-	}
-
-	void DiplomacyCard::play()
-	{
-		cout << "Playing DiplomacyCard" << endl;
-	}
-}
+//#include "Card.h"
+//
+//namespace Cards {
+//
+//	Card::Card() {
+//		this->name = new string("Basic Card");
+//	}
+//
+//	Card::~Card()
+//	{
+//		delete name;
+//	}
+//	
+//	void Card::play()
+//	{
+//		cout << "this is a basic Card" <<endl;
+//	}
+//
+//	string Card::getName()
+//	{
+//		return *name;
+//	}
+//
+//	void Card::print()
+//	{
+//		cout << this->name;
+//	}
+//
+//
+//	SpyCard::SpyCard()
+//	{
+//		this->name= new std:: string("Spy Card");
+//	}
+//
+//	SpyCard::~SpyCard()
+//	{
+//	}
+//
+//	void SpyCard::play()
+//	{
+//		cout << "Playing SpyCard" << endl;
+//	}
+//
+//	BombCard::BombCard()
+//	{
+//		this->name = new std::string("BombCard");
+//	}
+//
+//	BombCard::~BombCard()
+//	{
+//	}
+//
+//	void BombCard::play()
+//	{
+//		cout << "Playing BombCard" << endl;
+//	}
+//
+//	ReinforcementCard::ReinforcementCard()
+//	{
+//		name = new std::string("ReinforcementCard");
+//	}
+//
+//	ReinforcementCard::~ReinforcementCard()
+//	{
+//	}
+//
+//	void ReinforcementCard::play()
+//	{
+//		cout << "Playing ReinforcementCard" << endl;
+//	}
+//
+//	BlockadeCard::BlockadeCard()
+//	{
+//		name = new std::string("BlockadeCard");
+//	}
+//
+//	BlockadeCard::~BlockadeCard()
+//	{
+//		
+//	}
+//
+//	void BlockadeCard::play()
+//	{
+//		cout << "Playing BlockadeCard" << endl;
+//	}
+//		
+//	AirliftCard::AirliftCard()
+//	{
+//		name = new std::string("AirliftCard");
+//	}
+//
+//	AirliftCard::~AirliftCard()
+//	{
+//		
+//	}
+//
+//	void AirliftCard::play()
+//	{
+//		cout << "Playing AirliftCard" << endl;
+//	}
+//
+//	DiplomacyCard::DiplomacyCard()
+//	{
+//		name = new std::string("DiplomacyCard");
+//	}
+//
+//	DiplomacyCard::~DiplomacyCard()
+//	{
+//		
+//	}
+//
+//	void DiplomacyCard::play()
+//	{
+//		cout << "Playing DiplomacyCard" << endl;
+//	}
+//}
