+++ warzone/Hand.h
@@ -1,25 +1,25 @@
-#pragma once
-#include <vector>
-#include "Deck.h"
-
-using namespace std;
-
-namespace Cards {
-
-	class Hand
-	{
-	private:
-		Deck* deck;
-		vector<int*>* currentHand;
-
-	public:
-		Hand(Deck* deck);
-		void addCard(int* cardId);
-		void play(int index);
-
-		void printHand();
-
-	};
-
-}
-
+//#pragma once
+//#include <vector>
+//#include "Deck.h"
+//
+//using namespace std;
+//
+//namespace Cards {
+//
+//	class Hand
+//	{
+//	private:
+//		Deck* deck;
+//		vector<int*>* currentHand;
+//
+//	public:
+//		Hand(Deck* deck);
+//		void addCard(int* cardId);
+//		void play(int index);
+//
+//		void printHand();
+//
+//	};
+//
+//}
+//
