+++ warzone/Main.h
@@ -2,4 +2,3 @@
 #include "CardDriver.h"
 
 int main();
-
